scripts/requirements.txt
future==0.18.3
httplib2==0.22.0
nebula3-python==3.4.0
numpy==1.24.2
pandas==2.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
